Added fps counter and faith bar.

+++ Assets/Scripts/Powers/Power.cs
@@ -5,7 +5,7 @@ namespace Powers
 { 
     public abstract class Power : ScriptableObject
     {
-        public float FaithCost;
+        public int FaithCost;
         public float TotalCoolDown;
 
         public abstract void Invoke(TerrainModification terrainModification);
